% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importbamreadcount.R
\name{importbamreadcount}
\alias{importbamreadcount}
\title{Import bam-readcount output as a tibble}
\usage{
importbamreadcount(infile, warnings = FALSE)
}
\arguments{
\item{infile}{bam-readcount tab-delimited file to import}

\item{warnings}{show warnings}
}
\value{
A tibble
}
\description{
\code{importbamreadcount()} reads a bam-readcount tab-delimited file and saves
the contents in a tibble.
}
\details{
Columns:
\enumerate{
\item Chromosome
\item Position
\item Reference base
\item Depth
\item Stats for base \code{=}
\item Stats for base \code{A}
\item Stats for base \code{C}
\item Stats for base \code{G}
\item Stats for base \code{T}
\item Stats for base \code{N}
\item Stats for indel
}

Each reported base/indel contains a set of \code{:}-separated fields with the following statistics:\tabular{ll}{
   Field \tab Description \cr
   base \tab The base, eg \code{C} \cr
   count \tab Number of reads \cr
   avg_mapping_quality \tab Mean mapping quality \cr
   avg_basequality \tab Mean base quality \cr
   avg_se_mapping_quality \tab Mean single ended mapping quality \cr
   num_plus_strand \tab Number of reads on the plus/forward strand \cr
   num_minus_strand \tab Number of reads on the minus/reverse strand \cr
   avg_pos_as_fraction \tab Average position on the read as a fraction, calculated with respect to the length after clipping. This value is normalized to the center of the read: bases occurring strictly at the center of the read have a value of 1 \cr
   , those occurring strictly at the ends should approach a value of 0 \tab  \cr
   avg_num_mismatches_as_fraction \tab Average number of mismatches on these reads per base \cr
   avg_sum_mismatch_qualities \tab Average sum of the base qualities of mismatches in the reads \cr
   num_q2_containing_reads \tab Number of reads with q2 runs at the 3’ end \cr
   avg_distance_to_q2_start_in_q2_reads \tab Average distance of position (as fraction of unclipped read length) to the start of the q2 run \cr
   avg_clipped_length \tab Average clipped read length \cr
   avg_distance_to_effective_3p_end \tab Average distance to the 3’ prime end of the read (as fraction of unclipped read length) \cr
}
}
\examples{
importbamreadcount(system.file("extdata", "hg38.bwa.bamreadcount.tsv.gz", package = "importbio"))
}
