% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importbed.R
\name{importbed}
\alias{importbed}
\title{Import BED as a tibble}
\usage{
importbed(infile)
}
\arguments{
\item{infile}{BED file to import}
}
\value{
A tibble
}
\description{
Import BED as a tibble
}
\examples{
importbed(system.file("extdata", "refseq.bed.gz", package = "importbio"))
}
