% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgaf.R
\name{importgaf}
\alias{importgaf}
\title{Import a GAF as a tibble}
\usage{
importgaf(infile)
}
\arguments{
\item{infile}{GAF file to import}
}
\value{
A tibble
}
\description{
\code{importgaf()} reads a version 2.2 GO Annotation File (GAF) file.
}
\details{
Columns:
\itemize{
\item \code{DB}: Refers to the database from which the identifier in DB object ID (column 2) is drawn.
\item \code{DB_Object_ID}: A unique identifier from the database in DB (column 1) for the item being annotated.
\item \code{DB_Object_Symbol}: A (unique and valid) symbol to which DB object ID is matched.
\item \code{Relation}: Describes how a gene product relates to the GO term with which it is associated.
\item \code{GO_ID}: The GO identifier for the term attributed to the DB object ID.
\item \code{DB_Reference}: One or more unique identifiers for a single source cited as an authority for the attribution of the GO ID to the DB Object ID.
\item \code{Evidence_Code}: Evidence codes for GO annotations.
\item \code{With_From}: This field is used to hold an additional identifier for annotations.
\item \code{Aspect}: Refers to the namespace or ontology to which the GO ID (column 5) belongs.
\item \code{DB_Object_Name}: Name of gene or gene product.
\item \code{DB_Object_Synonym}: Gene symbol or other text.
\item \code{DB_Object_Type}: A description of the type of gene product being annotated.
\item \code{Taxon}: Taxonomic identifier(s).
\item \code{Date}: Date on which the annotation was made; format is YYYYMMDD.
\item \code{Assigned_By}: The database which made the annotation.
\item \code{Annotation_Extension}: Contains cross references to other ontologies that can be used to qualify or enhance the annotation.
\item \code{Gene_Product_Form_ID}: This field allows the annotation of specific variants of that gene or gene product.
}
}
\examples{
importgaf(system.file("extdata", "GCF_043380555.1-RS_2024_12_gene_ontology.gaf.gz", package = "importbio"))
}
