% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgff.R
\name{importgff}
\alias{importgff}
\title{Import GFF as a tibble}
\usage{
importgff(infile, sep = "=")
}
\arguments{
\item{infile}{GFF file to import}

\item{sep}{separator for group field}
}
\value{
A tibble
}
\description{
Import GFF as a tibble
}
\examples{
importgff(system.file("extdata", "gencode.v38.annotation.sample.gff3.gz", package = "importbio"))
}
