% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importmhcflurry.R
\name{importmhcflurry}
\alias{importmhcflurry}
\title{Import mhcflurry-predict output as a tibble}
\usage{
importmhcflurry(infile)
}
\arguments{
\item{infile}{\code{mhcflurry-predict} csv-delimited file to import}
}
\value{
A tibble
}
\description{
\code{importmhcflurry()} reads the CSV predictions from \code{mhcflurry-predict} and saves the contents in a tibble.
}
\details{
Columns:
\itemize{
\item \code{id}: identifier.
\item \code{allele}: HLA allele.
\item \code{peptide}: peptide sequence.
\item \code{mhcflurry_affinity}: binding affinity predictions as affinities (KD) in nM; lower values indicate stronger binders and a commonly-used threshold for peptides with a reasonable chance of being immunogenic is 500 nM.
\item \code{mhcflurry_affinity_percentile}: the percentile of the affinity prediction among a large number of random peptides tested on that allele (range 0 - 100). Lower is stronger. Two percent is a commonly-used threshold.
\item \code{mhcflurry_processing_score}: antigen processing score, which range from 0 to 1 with higher values indicating more favourable processing. The processing predictor is experimental.
\item \code{mhcflurry_presentation_score}: antigen presentation score, which range from 0 to 1 with higher values indicating more favourable presentation.
\item \code{mhcflurry_presentation_percentile}: this is not officially documented; should be the percentile of the presentation score among a large number of random peptides.
}
}
\examples{
importmhcflurry(system.file("extdata", "mhcflurry.pred.csv.gz", package = "importbio"))
}
