% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importvcf.R
\name{importvcf}
\alias{importvcf}
\title{Import VCF as tibble}
\usage{
importvcf(infile)
}
\arguments{
\item{infile}{VCF to import}
}
\value{
A tibble
}
\description{
Import VCF as tibble
}
\examples{
importvcf(system.file("extdata", "Pfeiffer.vcf.gz", package = "importbio"))
}
