% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importbamreadcount.R
\name{parse_base}
\alias{parse_base}
\title{Parse base strings}
\usage{
parse_base(x)
}
\arguments{
\item{x}{input string containing read stats}
}
\value{
A tibble
}
\description{
Meant for internal use only.
}
