% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skip_lines}
\alias{skip_lines}
\title{Helper functions}
\usage{
skip_lines(x, regex = "^#")
}
\arguments{
\item{x}{An input vector}

\item{regex}{A regular expression}
}
\value{
A vector
}
\description{
\code{skip_line()} skips line/s that match a regex
}
\examples{
skip_lines(1:10, 1)
skip_lines(c('the', 'quick', 'thread', 'THREE'), '^th')
}
