% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importbed.R
\name{importbed}
\alias{importbed}
\title{Import BED as a tibble}
\usage{
importbed(infile)
}
\arguments{
\item{infile}{BED file to import}
}
\value{
A tibble
}
\description{
\code{importbed()} reads a BED file and saves the contents in a tibble. The BED file must have at least three columns.
}
\details{
Columns:
\itemize{
\item \code{chrom}: The name of the chromosome (e.g. chr3, chrY, chr2_random) or scaffold (e.g. scaffold10671).
\item \code{chromStart}: The starting position of the feature in the chromosome or scaffold. The first base in a chromosome is numbered 0.
\item \code{chromEnd}: The ending position of the feature in the chromosome or scaffold. The chromEnd base is not included in the display of the feature, however, the number in position format will be represented. For example, the first 100 bases of chromosome 1 are defined as chrom=1, chromStart=0, chromEnd=100, and span the bases numbered 0-99 in our software (not 0-100), but will represent the position notation chr1:1-100. Read more here.
\item \code{name}: Defines the name of the BED line. This label is displayed to the left of the BED line in the Genome Browser window when the track is open to full display mode or directly to the left of the item in pack mode.
\item \code{score}: A score between 0 and 1000. If the track line useScore attribute is set to 1 for this annotation data set, the score value will determine the level of gray in which this feature is displayed (higher numbers = darker gray). This table shows the Genome Browser's translation of BED score values into shades of gray:
\item \code{strand}: Defines the strand. Either "." (=no strand) or "+" or "-".
\item \code{thickStart}: The starting position at which the feature is drawn thickly (for example, the start codon in gene displays). When there is no thick part, thickStart and thickEnd are usually set to the chromStart position.
\item \code{thickEnd}: The ending position at which the feature is drawn thickly (for example the stop codon in gene displays).
\item \code{itemRgb}: An RGB value of the form R,G,B (e.g. 255,0,0). If the track line itemRgb attribute is set to "On", this RBG value will determine the display color of the data contained in this BED line. NOTE: It is recommended that a simple color scheme (eight colors or less) be used with this attribute to avoid overwhelming the color resources of the Genome Browser and your Internet browser.
\item \code{blockCount}: The number of blocks (exons) in the BED line.
\item \code{blockSizes}: A comma-separated list of the block sizes. The number of items in this list should correspond to blockCount.
\item \code{blockStarts}: A comma-separated list of block starts. All of the blockStart positions should be calculated relative to chromStart. The number of items in this list should correspond to blockCount.
}
}
\examples{
importbed(system.file("extdata", "refseq.bed.gz", package = "importbio"))
}
