% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importblast.R
\name{importblast}
\alias{importblast}
\title{Import BLAST tabular output as a tibble}
\usage{
importblast(infile)
}
\arguments{
\item{infile}{BLAST tab-delimited file to import}
}
\value{
A tibble
}
\description{
\code{importblast()} reads a BLAST tab-delimited file (output format 6 or 7, where lines starting with a \verb{#} are ignored) and saves the contents in a tibble.
}
\examples{
importblast(system.file("extdata", "blastout.tsv.gz", package = "importbio"))
}
