% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importgff.R
\name{importgff}
\alias{importgff}
\title{Import GFF as a tibble}
\usage{
importgff(infile, sep = "=")
}
\arguments{
\item{infile}{GFF file to import}

\item{sep}{separator for group field}
}
\value{
A tibble
}
\description{
\code{importgff()} reads GFF output and saves the contents in a tibble.
}
\details{
Columns:
\itemize{
\item \code{seqname}: The name of the sequence; must be a chromosome or scaffold.
\item \code{source}: The program that generated this feature.
\item \code{feature}: The name of this type of feature, e.g. "CDS", "start_codon", "stop_codon", and "exon"
\item \code{start}: The starting position of the feature in the sequence; the first base is numbered 1.
\item \code{end}: The ending position of the feature (inclusive).
\item \code{score}: A score between 0 and 1000.
\item \code{strand}: Valid entries include "+", "-", or ".".
\item \code{frame}: If the feature is a coding exon, frame should be a number between 0-2 that represents the reading frame of the first base. If the feature is not a coding exon, the value should be ".".
\item \code{group}: All lines with the same group are linked together into a single item.
}
}
\examples{
importgff(system.file("extdata", "gencode.v38.annotation.sample.gff3.gz", package = "importbio"))
}
