% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importvcf.R
\name{importvcf}
\alias{importvcf}
\title{Import VCF as tibble}
\usage{
importvcf(infile)
}
\arguments{
\item{infile}{VCF to import}
}
\value{
A tibble
}
\description{
\code{importvcf()} reads VCF output and saves the contents in a tibble.
}
\details{
See \href{https://github.com/davetang/learning_vcf_file}{Learning the Variant Call Format} for more information.
}
\examples{
importvcf(system.file("extdata", "Pfeiffer.vcf.gz", package = "importbio"))
}
